finalizando us08

+++ src/br/edu/ufcg/computacao/p2lp2/hotelcalifornia/pagamento/DinheiroPagamento.java
@@ -1,7 +1,7 @@
 package br.edu.ufcg.computacao.p2lp2.hotelcalifornia.pagamento;
 
 public class DinheiroPagamento extends Pagamento{
-	public DinheiroPagamento(int idReserva, String idCliente, double valorEfetivamentePago, String nomeTitular) {
+	public DinheiroPagamento(long idReserva, String idCliente, double valorEfetivamentePago, String nomeTitular) {
         super(idReserva, idCliente, "Dinheiro", valorEfetivamentePago, nomeTitular);
     }
 }
